[SPARK-1986][GraphX]move lib.Analytics to org.apache.spark.examples
to support ~/spark/bin/run-example GraphXAnalytics triangles
/soc-LiveJournal1.txt --numEPart=256

move to graphx package, correct reference in LiveJournalPageRank

can now run:
~/spark/bin/run-example graphx.Analytics triangles soc-LiveJournal1.txt
--numEPart=256
~/spark/bin/run-example graphx.LiveJournalPageRank
soc-LiveJournal1.txt --numEPart=256
